Most changes due to exporting program changes
Most changes due to exporting program changes

There is a recent update to fix some issues with displaying and exporting, hence the program developer has issued a patch/update that results in large number of files being affected.

The changes may also be due to Discord display changes (CSS, HTML changes)

There may be some minor details not spotted, which is not the changes due to the software update.