Known username changes: Xzester#4520 -> trippydouche#4520
Xzester#4520 -> trippydouche#4520 was previously user and contributor of "study group a" server, a.k.a. a cheater. Please view history of commits on this repository (during June) for more information, because repository tracks changes from June, this user has DELETED messages from server "study group a" in order to hide or conceal information of cheating.